Added optimization capabilities.
Interface is kind of clunky still but running test without wire_lengths (or with an empty dictionary) will cause it to try to minimize the total manhattan distance. Unfortunately, because of the representation it is much slower for fabric dimensions that are not powers of two. So small_test is now on an 8x8 fabric.