Made small changes to work with the global router
Can run global_router.test(filepath) to place and (globally) route a .dot file. Currently just prints satisfying paths as a list of nodes. Added placer class for use with global_router. Does not yet backtrack in the case of unroutable nodes (although it should never be unroutable because it's not considering overlap yet)