Relaxed distance constraints for registers and retrying if unroutable
Because of the register placements, sometimes they need to take a strange path,
so I don't add distance constraints for registers. Note: This is generally fine
because if the registers are directly routable then monosat tends to take the
easy path.

Additionally, there *are* possible placements that are completely unroutable.
I have an example of that if you're interested.

These cases seem to be rare in practice. So to handle this for now, the place
and route part of test is in a while loop that tries up to 10 times to find
a valid placement and route.