Testing multiple solvers
Created a bash script in util to automatically iterate through different
placement solvers.

This also included restructuring test.py and pnr.py to allow passing
different solvers.

I had to instantiate pnr earlier in test.py to access the instantiated
solver for use in POSITION_T

I used try/catch with an eval to go from the solver string to the actual
solver. If you have a preference for a different way of doing that, just
let me know! Or feel free to change it yourself.

Also, I fixed a bug in the i/o ports when building the fabric.