Adding pe and mem tile locations. Routing currently doesn't work!
fab.pe_locations is a dictionary with
True: set of positions with pe's
False: set of positions without pe's

fab.mem_locations is a set of all the memory tile locations
This means the top of the memory tile (where the ports are) but NOT
the other rows of that memory tile. All that information is obtainable
from the pe_locations (at least until there are more types of tiles)

Let me know if you want a different interface.

I'm not currently parsing the new fabric completely, so routing doesn't work
In fact even parsing it throws an error, so just use --noroute