Support registers
Cleaning up and fixing interface with bitstream writer

Everything is functioning correctly now. Although there are a few notes

1. build_net_graphs is still unsupported and has not been ported to the
   new format yet
2. the bitstream writer interface has not been changed much. Soon, we
   should start relying more on the configuration engine
3. Occasionally the runtime is very slow -- I'm worried there's some bug
   in the fabric builder but I haven't found anything yet
   memory tiles seem to be connected correctly which is typically the issue
   -- I will look into this further. It might have to do with splitting
      register ports

Other notes:
I've reorganized a bit. I renamed fabricfuns to fabricutils. Additionally,
I put most of the fabric parsing code into pnr. My reasoning here is that
the fabric itself shouldn't know anything about the xml. Pnr (or possibly
another directory to be created) should know about this and *produces* a fabric.

As part of this, I created a pnrutils module for useful functions used in pnr. I
did not put this in the utils directory because it is specific to fabric parsing
as opposed to a general class/data structure.

I didn't end up using it, but I also created a DefaultSelectDict which does
exactly what you would expect it to -- although the default value has to be
hashable which is a bit limiting (why I didn't use it)

Finally, I made a PNRConfig class. I'm not attached to that structure. At some point
we should make a better configuration engine. I'm also thinking it could be instantiated
in pnr instead of externally, but because of the resets it's easier to have it separate.
In a future commmit I might make it a singleton.