Add an extra constraint for the recommended route settings
Because registers are not split in this encoding, we need
to enforce unreachability constraints to some other modules

This is to prevent the following situation:
a --> r1 --> b
|
-->c

without these new constraints, a-->r1-->c would be a potential
Monosat output.

The hope is that only enforcing not-reaches constraints on registers
and the few relevant modules is less expensive than using them for every
pair of unconnected modules.